data-pipeline/requirements.txt
apache-airflow==2.8.0
pandas==2.1.0
numpy==1.26.0
requests==2.31.0
matplotlib==3.8.0
seaborn==0.13.0
dvc==3.38.0
pytest==7.4.0